Serialize NaN and Infinity values as strings (#1322)
Float values that are "out of range" were previously serialized—to send to the
tracker/Hamilton UI—in a way that produced invalid JSON (but value JS), which
caused a runtime error trying to send values to the tracker.

`math.nan`, `math.inf`, and `-math.inf` are serialized by `json.dumps()` to
`NaN`, `Infinity`, and `-Infinity` values by default(NOT as strings).

This commit uses `json.dumps(…, allow_nan=False)`, which raises a `ValueError`
for such values, and in that case converts them to strings.

The result is that Hamliton UI will show such values as **strings**
`"nan"`, `"inf"`, and `"-inf"`.
This could be confusing only to users who are also using such string values.

Fixes #1310

Co-authored-by: Gavin Kistner <gkistner@nvidia.com>